requirements.txt
streamlit
streamlit_code_editor
numpy
os-sys